+++ src/main/java/ru/kpfu/itis/model/Deck.java
@@ -4,13 +4,13 @@ import java.util.ArrayList;
 
 public class Deck {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private ArrayList<Card> cards;
     private ArrayList<Card> waste;
 
-    public Deck(Integer id, String name, String description, ArrayList<Card> cards) {
+    public Deck(Long id, String name, String description, ArrayList<Card> cards) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -21,11 +21,11 @@ public class Deck {
     public Deck() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
